+++ nodemon.json
@@ -1,0 +1,6 @@
+{
+  "ignore": [".git", "node_modules"],
+  "watch": ["src"],
+  "exec": "npm run sim",
+  "ext": "ts"
+}
