--- src/constants/index.js
@@ -1,4 +1,0 @@
-export const PORT = process.env.PORT || 3000;
-export const BCRYPT_SALT = process.env.BCRYPT_SALT || 10;
-export const SECRET_KEY = process.env.SECRET_KEY || 'secret';
-export const ACCESS_TOKEN_DURATION = process.env.ACCESS_TOKEN_DURATION || '15m';
