+++ src/constants/index.ts
@@ -1,0 +1,4 @@
+export const PORT: number | string = process.env.PORT || 3000;
+export const BCRYPT_SALT: number | string = process.env.BCRYPT_SALT || 10;
+export const SECRET_KEY: string = process.env.SECRET_KEY || 'secret';
+export const ACCESS_TOKEN_DURATION: string = process.env.ACCESS_TOKEN_DURATION || '15m';
