+++ src/index.ts
@@ -1,0 +1,28 @@
+import  express from 'express';
+import cors from 'cors';
+import bodyParser from 'body-parser';
+import compression from 'compression';
+import path from 'path';
+
+import routes from './routes/index';
+import { PORT } from './constants/index';
+
+const app = express();
+const appRouter = express.Router();
+
+app.use(bodyParser.urlencoded({ extended: false }));
+app.use(bodyParser.json());
+app.use(cors());
+app.use(compression());
+
+appRouter.get('/', (req, res) => {
+  res.sendFile(path.join(`${__dirname}/../public/index.html`));
+});
+
+app.use('/', appRouter);
+app.use('/api', routes);
+
+app.listen(PORT, () => {
+  // eslint-disable-next-line no-console
+  console.log(`Server is running on ${PORT} port.`);
+});
